Added Basic Menu Functionality
Added a main menu and pause menu, with graphics. It has buttons that can be controlled with the mouse. When you hit a button on a menu, you need to lift the mouse for the next menu.

TODO: Create nicer main menu background with a title.

The game loads everythig at once at the begining for quick game state changes (for example a speedrunner could go to the main menu and back very quickly).

TODO: Add a splash screen.

+++ LeviathanPC/Cutscene.cpp
@@ -420,7 +420,7 @@ bool Cutscene::render (GPU_Target *screen, Arena **arena, Player *player, GameSt
 
 			this->font.getFont ()->drawBox (screen, Font::getRect_s (0.216f, 0.705f, 0.90133333333f, 0.25555555555f), this->current_text->text);
 
-			this->dialogue_timer -= Input::getDelta ();
+			this->dialogue_timer -= (float) Input::getDelta ();
 
 		} else {
